+++ service/proxy_getter/coderbusy.go
@@ -1,54 +1,48 @@
-package get_proxy
+package proxy_getter
 
 import (
-	"fmt"
-	"strings"
-
 	"github.com/PuerkitoBio/goquery"
 	"github.com/tongsq/go-lib/logger"
 	"github.com/tongsq/go-lib/request"
 	"proxy-collect/consts"
 	"proxy-collect/service/common"
+
+	"strings"
 )
 
-func NewGetProxyIp3366() *getProxyIp3366 {
-	return &getProxyIp3366{}
+func NewGetProxyCoderBusy() *getProxyCoderBusy {
+	return &getProxyCoderBusy{}
 }
 
-type getProxyIp3366 struct {
+type getProxyCoderBusy struct {
 }
 
-func (s *getProxyIp3366) GetUrlList() []string {
+func (s *getProxyCoderBusy) GetUrlList() []string {
 	list := []string{
-		"http://www.ip3366.net/free/?stype=1",
-		"http://www.ip3366.net/free/?stype=2",
-	}
-	for i := 2; i < 4; i++ {
-		list = append(list, fmt.Sprintf("http://www.ip3366.net/free/?stype=1&page=%d", i))
-		list = append(list, fmt.Sprintf("http://www.ip3366.net/free/?stype=2&page=%d", i))
+		"https://proxy.coderbusy.com/",
 	}
 	return list
 }
-func (s *getProxyIp3366) GetContentHtml(requestUrl string) string {
+func (s *getProxyCoderBusy) GetContentHtml(requestUrl string) string {
 	h := &request.RequestHeaderDto{
 		UserAgent:               consts.USER_AGENT,
 		UpgradeInsecureRequests: "1",
 	}
-
-	logger.Info("get proxy from ip3366", logger.Fields{"url": requestUrl})
+	logger.Info("get proxy from coderbusy", logger.Fields{"url": requestUrl})
 	data, err := request.WebGet(requestUrl, h, nil)
 	if err != nil || data == nil {
-		logger.Error("get proxy from ip3366 fail", logger.Fields{"err": err, "data": data})
+		logger.Error("get proxy from coderbusy fail", logger.Fields{"err": err, "data": data})
+		logger.Error("get proxy from coderbusy fail", logger.Fields{"err": err, "data": data})
 		return ""
 	}
 	return data.Body
 }
 
-func (s *getProxyIp3366) ParseHtml(body string) [][]string {
+func (s *getProxyCoderBusy) ParseHtml(body string) [][]string {
 
 	doc, err := goquery.NewDocumentFromReader(strings.NewReader(body))
 	if err != nil {
-		logger.Error("read fail", logger.Fields{"err": err})
+		logger.Error(consts.GO_QUERY_READ_ERROR, logger.Fields{"err": err})
 		return nil
 	}
 	var proxyList [][]string
@@ -56,7 +50,7 @@ func (s *getProxyIp3366) ParseHtml(body string) [][]string {
 		td := selection.ChildrenFiltered("td").First()
 		host := strings.TrimSpace(td.Text())
 		td2 := selection.ChildrenFiltered("td").Eq(1)
-		port := strings.TrimSpace(td2.Text())
+		port := strings.TrimSpace(td2.Find("a").First().Text())
 
 		if !common.CheckProxyFormat(host, port) {
 			logger.Error(consts.PROXY_FORMAT_ERROR, logger.Fields{"host": host, "port": port})
