+++ service/proxy_getter/list_plus.go
@@ -1,4 +1,4 @@
-package get_proxy
+package proxy_getter
 
 import (
 	"fmt"
@@ -11,55 +11,54 @@ import (
 	"proxy-collect/service/common"
 )
 
-func NewGetProxy89Ip() *getProxy89Ip {
-	return &getProxy89Ip{}
+func NewGetProxyProxyList() *getProxyProxyList {
+	return &getProxyProxyList{}
 }
 
-type getProxy89Ip struct {
+type getProxyProxyList struct {
 }
 
-func (s *getProxy89Ip) GetUrlList() []string {
+func (s *getProxyProxyList) GetUrlList() []string {
 	list := []string{
-		"https://www.89ip.cn/",
+		"https://list.proxylistplus.com/Fresh-HTTP-Proxy-List-1",
 	}
 	for i := 2; i < 6; i++ {
-		list = append(list, fmt.Sprintf("https://www.89ip.cn/index_%d.html", i))
+		list = append(list, fmt.Sprintf("https://list.proxylistplus.com/Fresh-HTTP-Proxy-List-%d", i))
 	}
 	return list
 }
-func (s *getProxy89Ip) GetContentHtml(requestUrl string) string {
+func (s *getProxyProxyList) GetContentHtml(requestUrl string) string {
 	h := &request.RequestHeaderDto{
 		UserAgent:               consts.USER_AGENT,
 		UpgradeInsecureRequests: "1",
-		Host:                    "www.89ip.cn",
-		Referer:                 "https://www.89ip.cn/",
+		Referer:                 "https://list.proxylistplus.com/update-2",
 	}
 
-	logger.Info("get proxy from 89ip", logger.Fields{"url": requestUrl})
+	logger.Info("get proxy from list.proxylistplus.com", logger.Fields{"url": requestUrl})
 	data, err := request.WebGet(requestUrl, h, nil)
 	if err != nil || data == nil {
-		logger.Error("get proxy from 89ip fail", logger.Fields{"err": err, "data": data})
+		logger.Error("get proxy from list.proxylistplus.com fail", logger.Fields{"err": err, "data": data})
 		return ""
 	}
 	return data.Body
 }
 
-func (s *getProxy89Ip) ParseHtml(body string) [][]string {
+func (s *getProxyProxyList) ParseHtml(body string) [][]string {
 
 	doc, err := goquery.NewDocumentFromReader(strings.NewReader(body))
 	if err != nil {
 		logger.Error(consts.GO_QUERY_READ_ERROR, logger.Fields{"err": err})
 		return nil
 	}
+
 	var proxyList [][]string
-	doc.Find("tbody > tr").Each(func(i int, selection *goquery.Selection) {
-		td := selection.ChildrenFiltered("td").First()
+	doc.Find("tr").Each(func(i int, selection *goquery.Selection) {
+		td := selection.ChildrenFiltered("td").Eq(1)
 		host := strings.TrimSpace(td.Text())
-		td2 := selection.ChildrenFiltered("td").Eq(1)
+		td2 := selection.ChildrenFiltered("td").Eq(2)
 		port := strings.TrimSpace(td2.Text())
 
 		if !common.CheckProxyFormat(host, port) {
-			logger.Error(consts.PROXY_FORMAT_ERROR, logger.Fields{"host": host, "port": port})
 			return
 		}
 		proxyArr := []string{host, port}
