+++ service/proxy_getter/proxy_7yip.go
@@ -1,4 +1,4 @@
-package get_proxy
+package proxy_getter
 
 import (
 	"fmt"
@@ -11,54 +11,50 @@ import (
 	"proxy-collect/service/common"
 )
 
-func NewGetProxyFanQie() *getProxyFanQie {
-	return &getProxyFanQie{}
+func NewGetProxy7Yip() *getProxy7Yip {
+	return &getProxy7Yip{}
 }
 
-type getProxyFanQie struct {
+type getProxy7Yip struct {
 }
 
-func (s *getProxyFanQie) GetUrlList() []string {
+func (s *getProxy7Yip) GetUrlList() []string {
 	list := []string{
-		"https://www.fanqieip.com/free",
+		"https://www.7yip.cn/free/",
 	}
 	for i := 2; i < 6; i++ {
-		list = append(list, fmt.Sprintf("https://www.fanqieip.com/free/%d", i))
+		list = append(list, fmt.Sprintf("https://www.7yip.cn/free/?action=china&page=%d", i))
 	}
 	return list
 }
-
-func (s *getProxyFanQie) GetContentHtml(requestUrl string) string {
+func (s *getProxy7Yip) GetContentHtml(requestUrl string) string {
 	h := &request.RequestHeaderDto{
 		UserAgent:               consts.USER_AGENT,
-		Host:                    "www.fanqieip.com",
 		UpgradeInsecureRequests: "1",
-		Referer:                 "https://www.fanqieip.com/free",
+		Host:                    "www.7yip.cn",
+		Referer:                 "https://www.7yip.cn/",
 	}
-
-	logger.Info("get proxy from fanqieip", logger.Fields{"url": requestUrl})
+	logger.Info("get proxy from 7yip", logger.Fields{"url": requestUrl})
 	data, err := request.WebGet(requestUrl, h, nil)
 	if err != nil || data == nil {
-		logger.Error("get proxy from fanqieip fail", logger.Fields{"err": err, "data": data})
-		return ""
+		logger.Error("get proxy from 7yip fail", logger.Fields{"err": err, "data": data})
 	}
 	return data.Body
 }
 
-func (s *getProxyFanQie) ParseHtml(body string) [][]string {
+func (s *getProxy7Yip) ParseHtml(body string) [][]string {
 
 	doc, err := goquery.NewDocumentFromReader(strings.NewReader(body))
 	if err != nil {
 		logger.Error(consts.GO_QUERY_READ_ERROR, logger.Fields{"err": err})
 		return nil
 	}
-
 	var proxyList [][]string
 	doc.Find("tbody > tr").Each(func(i int, selection *goquery.Selection) {
 		td := selection.ChildrenFiltered("td").First()
-		host := strings.TrimSpace(td.ChildrenFiltered("div").First().Text())
+		host := strings.TrimSpace(td.Text())
 		td2 := selection.ChildrenFiltered("td").Eq(1)
-		port := strings.TrimSpace(td2.ChildrenFiltered("div").First().Text())
+		port := strings.TrimSpace(td2.Text())
 
 		if !common.CheckProxyFormat(host, port) {
 			logger.Error(consts.PROXY_FORMAT_ERROR, logger.Fields{"host": host, "port": port})
@@ -67,5 +63,13 @@ func (s *getProxyFanQie) ParseHtml(body string) [][]string {
 		proxyArr := []string{host, port}
 		proxyList = append(proxyList, proxyArr)
 	})
+
 	return proxyList
 }
+
+//func (s *getProxy7Yip) GetSource() string {
+//	_, file, _, _ := runtime.Caller(0)
+//	arr := strings.Split(file, "/")
+//	name := arr[len(arr)-1]
+//	return name[0:len(name)-3]
+//}
