+++ service/proxy_getter/xici.go
@@ -1,4 +1,4 @@
-package get_proxy
+package proxy_getter
 
 import (
 	"fmt"
@@ -10,41 +10,42 @@ import (
 	"proxy-collect/consts"
 )
 
-func NewGetProxyXila() *getProxyXila {
-	return &getProxyXila{}
+func NewGetProxyXici() *getProxyXici {
+	return &getProxyXici{}
 }
 
-type getProxyXila struct {
+type getProxyXici struct {
 }
 
-func (s *getProxyXila) GetUrlList() []string {
+func (s *getProxyXici) GetUrlList() []string {
 	list := []string{
-		"http://www.xiladaili.com/https/",
+		"http://www.xicidaili.com/wn/",
 	}
 	for i := 2; i < 6; i++ {
-		list = append(list, fmt.Sprintf("http://www.xiladaili.com/https/%d/", i))
+		list = append(list, fmt.Sprintf("http://www.xicidaili.com/wn/%d", i))
 	}
 	return list
 }
 
-func (s *getProxyXila) GetContentHtml(requestUrl string) string {
+func (s *getProxyXici) GetContentHtml(requestUrl string) string {
 
 	h := &request.RequestHeaderDto{
 		UserAgent:               consts.USER_AGENT,
-		Host:                    "www.xiladaili.com",
-		Referer:                 "http://www.xiladaili.com/https/",
+		Host:                    "www.xicidaili.com",
+		Referer:                 "http://www.xicidaili.com",
 		UpgradeInsecureRequests: "1",
 	}
-	logger.Info("get proxy from xiladaili", logger.Fields{"url": requestUrl})
+
+	logger.Info("get proxy from xicidaili", logger.Fields{"url": requestUrl})
 	data, err := request.WebGet(requestUrl, h, nil)
 	if err != nil || data == nil {
-		logger.Error("get proxy from xiladaili fail", logger.Fields{"err": err, "data": data})
+		logger.Error("get proxy from xicidaili fail", logger.Fields{"err": err, "data": data})
 		return ""
 	}
 	return data.Body
 }
 
-func (s *getProxyXila) ParseHtml(body string) [][]string {
+func (s *getProxyXici) ParseHtml(body string) [][]string {
 	doc, err := goquery.NewDocumentFromReader(strings.NewReader(body))
 	if err != nil {
 		logger.Error(consts.GO_QUERY_READ_ERROR, logger.Fields{"err": err})
