+++ service/proxy_getter/zdaye_index.go
@@ -1,4 +1,4 @@
-package get_proxy
+package proxy_getter
 
 import (
 	"regexp"
